execution/state: move the BAL benchmarks to their own file
versionedio_test.go carries the tests; the package already keeps benchmark-only
fixtures in *_bench_test.go.
